more performance updates
